(PC-13461) feat(SetBirthday): add tests for DatePickerSpinner.native
