+++ src/features/auth/signup/SetBirthday/SetBirthday.web.test.tsx
@@ -1,8 +1,7 @@
 import mockdate from 'mockdate'
 import React from 'react'
 
-import { analytics } from 'libs/analytics'
-import { fireEvent, render } from 'tests/utils/web'
+import { render } from 'tests/utils/web'
 
 import { SetBirthday } from './SetBirthday'
 
@@ -12,44 +11,26 @@ const props = { goToNextStep: jest.fn(), signUp: jest.fn() }
 
 jest.mock('features/auth/settings')
 
+// eslint-disable-next-line local-rules/no-allow-console
+allowConsole({ warn: true })
+
 describe('<SetBirthday />', () => {
   beforeEach(() => {
     mockdate.set(CURRENT_DATE)
     jest.useFakeTimers()
   })
 
-  describe('- No touch device -', () => {
-    it('should render correctly', () => {})
-
-    describe('- navigation -', () => {
-      it('should keep disabled the button "Continuer" when the date is not selected', () => {})
-
-      it('should keep enable the button "Continuer" when the date is selected and is different from the current date', () => {})
-
-      it('should call goToNextStep() when the date is selected and press the button "Continuer"', () => {})
+  describe('touch device', () => {
+    it('should render correctly', () => {
+      const renderAPI = render(<SetBirthday {...props} />, { theme: { isTouch: true } })
+      expect(renderAPI).toMatchSnapshot()
     })
+  })
 
-    describe('- error message -', () => {
-      it('should display the error message "tu dois avoir 15 ans" when the selected date is too young', () => {})
-
-      it('should not display the error message "tu dois avoir 15 ans" when the user is exactly 15yo', () => {})
-
-      it('should not display the error message "tu dois avoir 15 ans" when the selected date is the current date', () => {})
-    })
-
-    describe('- analytics -', () => {
-      it('should log ConsultModalWhyAnniversary when clicking "Pour quelle raison ?" button', () => {
-        const { getByText } = render(<SetBirthday {...props} />)
-
-        const whyBirthdayLink = getByText('Pour quelle raison ?')
-        fireEvent.click(whyBirthdayLink)
-
-        expect(analytics.logConsultWhyAnniversary).toHaveBeenCalledTimes(1)
-      })
-
-      it('should not log SignUpTooYoung if the user is 15 years old or more', () => {})
-
-      it('should log SignUpTooYoung if the user is 14 years old or less', () => {})
+  describe('no touch device', () => {
+    it('should render correctly', () => {
+      const renderAPI = render(<SetBirthday {...props} />, { theme: { isTouch: false } })
+      expect(renderAPI).toMatchSnapshot()
     })
   })
 })
